+++ CHANGELOG.md
@@ -81,6 +81,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [#7857](https://github.com/osmosis-labs/osmosis/pull/7857) SuperfluidDelegationsByValidatorDenom query now returns equivalent staked amount
 * [#7875](https://github.com/osmosis-labs/osmosis/pull/7875) Introduce Chain interface for the x/bridge
 * [#7880](https://github.com/osmosis-labs/osmosis/pull/7880) Rework x/bridge BTC component to use Chain interface
+* [#7882](https://github.com/osmosis-labs/osmosis/pull/7882) Add confirmations check for x/bridge transfers
 
 ### SDK